Simplify the error msg returned from the `read` method in cypher.go
When `cm-neo4j-driver` return an error in the `read` method in cypher.go,
the entire cypher statement is being written into the error and later logged by
the logger. The cypher statement is always the same and it doesn't add much
value to the error message, that's why it's being removed.