Lazy fix linter error - Exported function returns unexported type
`NewCypherDriver` function in `annotations` was returning `cypherDriver` type
and as a result golint wasn't happy. The solution was to export it.

The fix is lazy because the `CypherDriver` type has unexported fields and
methods and can't be used from outside the `annotations` package. To export it's
methods results in a cascading exports (the return types of the methods must be
exported as well) that's why the non-lazy fix will be applied in another commit
and another PR.

Also fix another linter error in the tests which is about unused result of a
function.